+++ src/cljs/camelot/component/library/core.cljs
@@ -136,7 +136,7 @@
                                                               [:selected-survey :survey-id :value] -1)})
       (om/update! data [:library :search :page] 1)
       (om/update! data [:library :survey-id] (get-in (state/app-state-cursor) [:selected-survey :survey-id :value]))
-      (om/update! data [:library :search :show-select-count] 0)
+      (om/update! data [:library :notification] nil)
       (om/update! data [:library :identification] {:quantity 1 :sighting-fields {}})
       (rest/get-x "/sighting-fields"
                   (fn [r]